Separate dropdown state and event handling into separate hook - RSC-449
